refactor: delete subpackages for exceptions and types
