+++ src/main/java/org/whdl/intermediate/EndpointType.java
@@ -1,6 +1,5 @@
-package org.whdl.intermediate.types;
+package org.whdl.intermediate;
 
-import org.whdl.intermediate.Type;
 
 public class EndpointType extends Type {
   private String definitionName;
