+++ src/test/java/org/whdl/intermediate/TestValue.java
@@ -3,8 +3,6 @@ package org.whdl.intermediate;
 import static org.junit.Assert.*;
 
 import org.junit.Test;
-import org.whdl.intermediate.exceptions.UndefinedBehaviourError;
-import org.whdl.intermediate.types.NodeType;
 
 public class TestValue {
